+++ apps/app/src/hooks/useProvider/index.ts
@@ -1,0 +1,1 @@
+export { useProvider } from "./useProvider";
